+++ test/fixtures/local-async-valid.template.js
@@ -1,0 +1,6 @@
+module.exports = function(opts, cb) {
+    cb(null, {
+        "Parameters": {},
+        "Resources": {}
+    });
+};